+++ src/controller/cargo.ts
@@ -2,30 +2,13 @@ import { async } from "@washanhanzi/result-enum"
 import { execAsync } from "@/util"
 import { CargoTomlTable, Ctx } from "@/entity"
 
-export async function cargoTree(ctx: Ctx) {
-	const tree = await async(execAsync(`cargo tree --manifest-path ${ctx.path} --depth 1 --all-features`))
+export async function cargoTree(path: string) {
+	const tree = await async(execAsync(`cargo tree --manifest-path ${path} --depth 1 --all-features`))
 	if (tree.isErr()) {
 		throw tree.unwrapErr()
 	}
 
 	return parseCargoTreeOutput(tree.unwrap() as string)
-}
-
-
-function parseCargoTree(output: string) {
-	const crateMap = {}
-	const lines = output.trim().split('\n')
-
-	lines.forEach(line => {
-		const trimmedLine = line.trim()
-		const match = trimmedLine.match(/(\S+)\s+v(\d+\.\d+\.\d+(?:[-+]\S+)?)/)
-		if (match) {
-			const [, crateName, version] = match
-			crateMap[crateName] = version
-		}
-	})
-
-	return crateMap
 }
 
 type CargoTreeOutputItem = {